Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/object-redirect-blocked.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
<meta http-equiv="Content-Security-Policy" content="object-src http://127.0.0.1:8000/resources/redirect.php">
</head>
<body>
<object type="text/html" data="http://127.0.0.1:8000/resources/redirect.php?code=307&url=http%3A%2F%2Flocalhost%3A8000/security/contentSecurityPolicy/resources/alert-fail.html"></object>
</body>
</html>
